fixed: factories.py, test_api_news.py
